프론트 연결하면서 고쳐야하는 부분 수정

+++ backend/src/main/java/com/maptist/mappride/mappride/photo/PhotoService.java
@@ -40,6 +40,9 @@ public class PhotoService {
         if(findPhoto.getMember() != memberService.getMember()){
             throw new RuntimeException("본인의 사진만 삭제할 수 있습니다.");
         }
+        if(findPhoto.isThumbnail()){
+            throw new RuntimeException("대표사진은 삭제할 수 없습니다.");
+        }
 
         s3Service.deleteFile(findPhoto.getPhotoUrl());
 
@@ -54,6 +57,7 @@ public class PhotoService {
 
         Member member = memberService.getMember();
         Place place = placeRepository.findOne(photoAddRequestDto.getPlaceId());
+        System.out.println(place);
 
         List<Long> photoIds = new ArrayList<>();
